Add maxItemsPerRow to limit number of items in a row
* Update index.d.ts

Add maxItemsPerRow to props in type definition

* Update utils.js

Add maxItemsPerRow to calculateDimensions function

* Update SectionGrid.js

Add maxItemsPerRow prop to SectionGrid

* Update FlatGrid.js

Add maxItemsPerRow prop to FlatGrid

* Add maxItemsPerRow to Properties doc

* Fix web onLayout - ignore when dimensions are 0

On web the onLayout event fires when navigating away from the screen and the dimensions are 0. This causes extra renders and flashes when navigating between screens.

* Fix web onLayout - ignore when dimensions are 0

On web the onLayout event fires when navigating away from the screen and the dimensions are 0. This causes extra renders and flashes when navigating between screens.