Use contentContainerStyle to adjust totalDimension (#185)
* Update index.d.ts

Add maxItemsPerRow to props in type definition

* Update utils.js

Add maxItemsPerRow to calculateDimensions function

* Update SectionGrid.js

Add maxItemsPerRow prop to SectionGrid

* Update FlatGrid.js

Add maxItemsPerRow prop to FlatGrid

* Add maxItemsPerRow to Properties doc

* Fix web onLayout - ignore when dimensions are 0

On web the onLayout event fires when navigating away from the screen and the dimensions are 0. This causes extra renders and flashes when navigating between screens.

* Fix web onLayout - ignore when dimensions are 0

On web the onLayout event fires when navigating away from the screen and the dimensions are 0. This causes extra renders and flashes when navigating between screens.

* defaultTotalDimension should not exceed screen dimensions

defaultTotalDimension should use same logic as onLocalLayout

* defaultTotalDimension should not exceed screen dimensions

defaultTotalDimension should use same logic as onLocalLayout

* Add adjustDimension function to utils.js

New function to adjust for max width or height and subtract padding from totalDimension

* Use new adjustDimension function in SectionGrid

* Use new adjustDimension function in FlatGrid

* Change " to '

* Added notes on how contentContainerStyle is used

* Also adjust totalDimension for Grid `style` prop padding

If padding is set in the `style` prop of the grid then the `totalDimension` should reflect that

* Pass `style` to `adjustDimension` call

* Pass `style` to `adjustDimension` call

* Update README.md

* Rename adjustDimension to getAdjustedTotalDimensions and add adjustGridToStyles prop

* Add adjustGridToStyles prop to types

* Rename function getAdjustedTotalDimensions + add adjustGridToStyles prop

* Rename function getAdjustedTotalDimensions + add adjustGridToStyles prop

* Update docs with adjustGridToStyles prop

* Update comment

* getAdjustedTotalDimensions newTotalDimension rename to totalDimension

* getAdjustedTotalDimensions newTotalDimension rename to totalDimension